+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/StreamFileCatalog.scala
@@ -23,8 +23,8 @@ import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SQLContext
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.execution.{FileCatalog, Partition}
 import org.apache.spark.sql.execution.datasources.PartitionSpec
-import org.apache.spark.sql.sources.{FileCatalog, Partition}
 import org.apache.spark.sql.types.StructType
 
 class StreamFileCatalog(sqlContext: SQLContext, path: Path) extends FileCatalog with Logging {
